possible transition of the permission tree node from a boolean granted state to a byte. This will allow for a new concept- * There are negations which would be defined as -1 * There are regular grants which would be defined as 1 * There are inherited which would be defined as 0.
If the given node is -1 or 1, it acts respectively to the rules above. If it is a 0, the node is effectively "removed" from the tree, which means that the tree has nether permission or no permission for said command. However, it would act like no permission when checked upon in the hasPermission method. This tri-state node makes it so we can have non transferable nodes across inheritance. So, when importing permissions from some other group, if a node is marked as inherit, it will not be transferred over because it is assumed to not have permission.